This PR introduces some optimisations to improve import times (and therefore startup times).

Initially in https://github.com/PyO3/maturin-import-hook/pull/7 I requested that the cache is able to be disabled entirely, but with `importlib.reload()` being an edge case and [`importlib.invalidate_caches()`](https://docs.python.org/3/library/importlib.html#importlib.invalidate_caches) existing it seems sufficient that the maturin import hook supports this way of clearing the cache to ensure new modules are detected.


# Benchmarking

Benchmarking was carried out using the `tests/create_benchmark_data.py` script. this script creates a scenario close to the worst case, with 100 paths in `sys.path` each very long nested paths with two `pyproject.toml`s present on the path back to the root of the filesystem that must be checked.

# Results:

These results are run on the slowest storage I have (a USB drive). Each value is the average over 3 runs.
The 'cold cache' runs are after running `echo 3 | sudo tee /proc/sys/vm/drop_caches`

|                                         | cold cache | warm cache |
| --                                      | --         | --         |
| `main`                                  | 8.80s      | 6.77s      |
| `is_maybe_maturin_project` optimisation | 7.36s      | 5.53s      |
| `lru_cache`                             | 1.95s      | 0.62s      |
| `find_dist_info_path` optimisation      | 1.69s      | 0.53s      |

(note that for a more reasonable case such as 20 paths in `sys.path`, the time taken is ~0.01s)